+++ src/main/kotlin/me/ruslanys/ifunny/service/Services.kt
@@ -1,0 +1,11 @@
+package me.ruslanys.ifunny.service
+
+import me.ruslanys.ifunny.channel.Channel
+import me.ruslanys.ifunny.channel.MemeInfo
+import me.ruslanys.ifunny.domain.Meme
+
+interface MemeService {
+
+    fun add(channel: Channel, info: MemeInfo): Meme
+
+}