+++ src/test/kotlin/me/ruslanys/ifunny/ApplicationTests.kt
@@ -1,0 +1,13 @@
+package me.ruslanys.ifunny
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class ApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+
+}